backend/requirements.txt
# Flask Framework
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
Flask-WTF==1.2.1
Flask-Caching==1.11.1

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
Flask-SQLAlchemy==3.1.1

# Authentication
PyJWT==2.8.0
bcrypt==4.1.2
itsdangerous==2.1.2

# Security
python-dotenv==1.0.0
marshmallow==3.20.1
webargs==9.0.0

# AI and NLP
openai==1.3.7
nltk==3.8.1
textstat==21.3.3
language-tool-python==1.5.0

# PDF Generation
WeasyPrint==60.1
pdfkit==1.0.0
ReportLab==4.0.8
xhtml2pdf==0.2.11

# DOCX Generation
python-docx==1.1.0

# Image Processing
Pillow==10.1.0
qrcode==7.4.2

# HTTP Requests
httpx==0.25.2
requests==2.31.0

# Utilities
python-dateutil==2.8.2
slugify==2.0.1
uuid==1.30

# Validation
email-validator==2.1.0
phone-validate==1.0.0

# Development
python-dotenv==1.0.0
watchdog==3.0.0

# Gunicorn for production
gunicorn==21.2.0

# Logging
colorlog==6.8.0
